Fix server action exports for property invites
