+++ app/property/invites/inviteCreateState.ts
@@ -1,0 +1,10 @@
+export type InviteCreateActionState = {
+  status: "idle" | "validation-error" | "invite-created";
+  message?: string;
+  warning?: string;
+  inviteLink?: string;
+  invitedEmail?: string;
+  expiresAt?: string;
+};
+
+export const initialInviteCreateActionState: InviteCreateActionState = { status: "idle" };
